Skiller ut logikk for å generere fritekstbrev html
