+++ src/main/kotlin/no/nav/familie/klage/brev/baks/BaksBrevOppretter.kt
@@ -5,12 +5,9 @@ import no.nav.familie.klage.behandling.BehandlingService
 import no.nav.familie.klage.behandling.domain.Behandling
 import no.nav.familie.klage.behandling.domain.StegType
 import no.nav.familie.klage.behandling.domain.erLåstForVidereBehandling
-import no.nav.familie.klage.brev.BrevClient
-import no.nav.familie.klage.brev.BrevsignaturService
 import no.nav.familie.klage.brev.FamilieDokumentClient
-import no.nav.familie.klage.fagsak.FagsakService
+import no.nav.familie.klage.felles.domain.Fil
 import no.nav.familie.klage.infrastruktur.exception.Feil
-import no.nav.familie.klage.personopplysninger.PersonopplysningerService
 import org.slf4j.LoggerFactory
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Component
@@ -18,51 +15,22 @@ import java.util.UUID
 
 @Component
 class BaksBrevOppretter(
-    private val brevClient: BrevClient,
     private val baksBrevRepository: BaksBrevRepository,
     private val behandlingService: BehandlingService,
     private val familieDokumentClient: FamilieDokumentClient,
-    private val brevsignaturService: BrevsignaturService,
-    private val fagsakService: FagsakService,
-    private val personopplysningerService: PersonopplysningerService,
-    private val fritekstBrevRequestDtoUtleder: FritekstBrevRequestDtoUtleder,
+    private val fritekstbrevHtmlUtleder: FritekstbrevHtmlUtleder,
 ) {
     private val logger = LoggerFactory.getLogger(BaksBrevOppretter::class.java)
 
     @Transactional
-    fun opprettBrev(behandlingId: UUID): ByteArray {
+    fun opprettBrev(behandlingId: UUID): BaksBrev {
         logger.debug("Oppretter brev for behandling {}", behandlingId)
-
         val behandling = behandlingService.hentBehandling(behandlingId)
         validerRedigerbarBehandling(behandling)
         validerKorrektBehandlingssteg(behandling)
-
-        val fagsak = fagsakService.hentFagsak(behandling.fagsakId)
-        val personopplysninger = personopplysningerService.hentPersonopplysninger(behandlingId)
-
-        val fritekstBrevRequestDto = fritekstBrevRequestDtoUtleder.utled(
-            fagsak,
-            behandling,
-            personopplysninger.navn,
-        )
-
-        val signaturDto = brevsignaturService.lagSignatur(
-            personopplysninger,
-            fagsak.fagsystem,
-        )
-
-        val fritekstbrevHtml = brevClient.genererHtmlFritekstbrev(
-            fritekstBrev = fritekstBrevRequestDto,
-            saksbehandlerNavn = signaturDto.navn,
-            enhet = signaturDto.enhet,
-        )
-
-        lagreEllerOppdaterBrev(
-            behandlingId = behandlingId,
-            html = fritekstbrevHtml,
-        )
-
-        return familieDokumentClient.genererPdfFraHtml(fritekstbrevHtml)
+        val fritekstbrevHtml = fritekstbrevHtmlUtleder.utledFritekstbrevHtml(behandling)
+        val pdfFraHtml = familieDokumentClient.genererPdfFraHtml(fritekstbrevHtml)
+        return lagreEllerOppdaterBrev(behandlingId, fritekstbrevHtml, pdfFraHtml)
     }
 
     private fun validerRedigerbarBehandling(behandling: Behandling) {
@@ -80,13 +48,14 @@ class BaksBrevOppretter(
     private fun lagreEllerOppdaterBrev(
         behandlingId: UUID,
         html: String,
+        pdf: ByteArray,
     ): BaksBrev {
         val brev = baksBrevRepository.findByIdOrNull(behandlingId)
         if (brev != null) {
-            val oppdatertBaksBrev = brev.copy(html = html)
+            val oppdatertBaksBrev = brev.copy(html = html, pdf = Fil(pdf))
             return baksBrevRepository.update(oppdatertBaksBrev)
         }
-        val nyttBaksBrev = BaksBrev(behandlingId = behandlingId, html = html)
+        val nyttBaksBrev = BaksBrev(behandlingId = behandlingId, html = html, pdf = Fil(pdf))
         return baksBrevRepository.insert(nyttBaksBrev)
     }
 }
